CustomerCareAssistant/requirements.txt
flask==2.1.1
openai==0.27.0
torch==1.10.0
numpy==1.21.0
scipy==1.7.1
pymupdf==1.18.10
pymongo==3.12.1
python-dotenv==0.19.2